refactor(fluent-icon): extract FluentIconType into a separate file
