+++ Desktop/Components/Controls/FluentIcon.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Reflection;
 using System.Windows;
+using StreamBoard.Core.Models;
 using Wpf.Ui.Controls;
 
 namespace StreamBoard.Components.Controls
@@ -8,48 +9,6 @@ namespace StreamBoard.Components.Controls
     public class IconGlyphAttribute(string glyph) : Attribute
     {
         public string Glyph { get; } = glyph;
-    }
-
-    public enum FluentIconType
-    {
-        [IconGlyph("\uE710")] Add,
-        [IconGlyph("\uED35")] Apps,
-        [IconGlyph("\uE74D")] Delete,
-        [IconGlyph("\uEA6C")] Devices,
-        [IconGlyph("\uE8A5")] Document,
-        [IconGlyph("\uE739")] Checkbox,
-        [IconGlyph("\uE70D")] ChevronDown,
-        [IconGlyph("\uE76C")] ChevronRight,
-        [IconGlyph("\uE8BB")] ChromeClose,
-        [IconGlyph("\uED62")] ClearAllInk,
-        [IconGlyph("\uE790")] Color,
-        [IconGlyph("\uE9A6")] FitPage,
-        [IconGlyph("\uE8B7")] Folder,
-        [IconGlyph("\uE12B")] Globe,
-        [IconGlyph("\uF0E2")] Grid,
-        [IconGlyph("\uE76F")] GripperBarHorizontal,
-        [IconGlyph("\uE897")] Help,
-        [IconGlyph("\uE80F")] Home,
-        [IconGlyph("\uE92E")] Keyboard,
-        [IconGlyph("\uE72E")] Lock,
-        [IconGlyph("\uE708")] Moon,
-        [IconGlyph("\uE74F")] Mute,
-        [IconGlyph("\uE968")] Network,
-        [IconGlyph("\uE91B")] Photo,
-        [IconGlyph("\uE7E8")] PowerButton,
-        [IconGlyph("\uEA86")] Puzzle,
-        [IconGlyph("\uE7C8")] Record,
-        [IconGlyph("\uF407")] RectangularClipping,
-        [IconGlyph("\uE8AC")] Rename,
-        [IconGlyph("\uE71E")] Search,
-        [IconGlyph("\uE713")] Settings,
-        [IconGlyph("\uE93E")] Streaming,
-        [IconGlyph("\uE71A")] Stop,
-        [IconGlyph("\uE895")] Sync,
-        [IconGlyph("\uE770")] System,
-        [IconGlyph("\uE916")] Timer,
-        [IconGlyph("\uE714")] Video,
-        [IconGlyph("\uE890")] View,
     }
 
     public partial class FluentIcon : FontIcon
