+++ Desktop/Features/Decks/Actions/Input/HotkeyAction.cs
@@ -1,9 +1,8 @@
 using System.Diagnostics;
 using System.Runtime.InteropServices;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using System.Windows.Input;
-using StreamBoard.Components.Controls;
+using StreamBoard.Core.Models;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
 using StreamBoard.Helpers;
