+++ Desktop/Features/Decks/Actions/Obs/ScreenshotAction.cs
@@ -1,10 +1,8 @@
-using System;
 using System.Diagnostics;
 using System.IO;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using StreamBoard.Components.Controls;
+using StreamBoard.Core.Models;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
 using StreamBoard.Features.Integrations.Obs.Services;
