+++ Desktop/Features/Decks/Actions/Obs/SwitchSceneAction.cs
@@ -1,13 +1,10 @@
-using System;
 using System.Diagnostics;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using Microsoft.Extensions.DependencyInjection;
-using StreamBoard.Core;
-using StreamBoard.Components.Controls;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
 using StreamBoard.Features.Integrations.Obs.Services;
+using StreamBoard.Core.Models;
 
 namespace StreamBoard.Features.Decks.Actions.Obs
 {
