+++ Desktop/Features/Decks/Actions/System/DelayAction.cs
@@ -1,5 +1,5 @@
 using System.Text.Json.Serialization;
-using StreamBoard.Components.Controls;
+using StreamBoard.Core.Models;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
 using StreamBoard.Helpers;
