+++ Desktop/Features/Decks/Actions/System/OpenFolderAction.cs
@@ -1,6 +1,6 @@
 using System.Diagnostics;
 using System.Text.Json.Serialization;
-using StreamBoard.Components.Controls;
+using StreamBoard.Core.Models;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
