+++ Desktop/Features/Decks/Services/ActionRegistry.cs
@@ -1,10 +1,9 @@
-﻿using StreamBoard.Components.Controls;
+﻿using StreamBoard.Core.Models;
 using StreamBoard.Features.Decks.Attributes;
 using StreamBoard.Features.Decks.Models;
 using StreamBoard.Features.Decks.ViewModels;
 using StreamBoard.Features.Integrations.Common.Models;
 using System.Reflection;
-using Wpf.Ui.Controls;
 
 namespace StreamBoard.Features.Decks.Services
 {
